fix: consolidate modelhost and service files for pkg-file-count limit
