+++ pkg/service/factory_build.go
@@ -1628,3 +1628,65 @@ func modelEventErrorClass(err error) string {
 	}
 	return "MODEL_EXECUTION_FAILED"
 }
+
+type zapModelHostLogger struct {
+	logger *zap.Logger
+}
+
+func newZapModelHostLogger(logger *zap.Logger) modelhost.Logger {
+	if logger == nil {
+		return nil
+	}
+	return zapModelHostLogger{logger: logger}
+}
+
+func (l zapModelHostLogger) Info(msg string, fields map[string]string) {
+	l.logger.Info(msg, modelHostZapFields(fields)...)
+}
+
+func (l zapModelHostLogger) Warn(msg string, fields map[string]string) {
+	l.logger.Warn(msg, modelHostZapFields(fields)...)
+}
+
+func modelHostZapFields(fields map[string]string) []zap.Field {
+	if len(fields) == 0 {
+		return nil
+	}
+	out := make([]zap.Field, 0, len(fields))
+	for key, value := range fields {
+		out = append(out, zap.String(key, value))
+	}
+	return out
+}
+
+type invocationMetricsRecorderAdapter struct {
+	recorder InvocationMetricsRecorder
+}
+
+func newModelHostMetricsRecorder(recorder InvocationMetricsRecorder) modelhost.MetricsRecorder {
+	if recorder == nil {
+		return nil
+	}
+	return invocationMetricsRecorderAdapter{recorder: recorder}
+}
+
+func (a invocationMetricsRecorderAdapter) RecordMetric(name string, labels map[string]string) {
+	if a.recorder == nil || strings.TrimSpace(name) == "" {
+		return
+	}
+	a.recorder.RecordInvocationMetric(InvocationMetric{
+		Name:   name,
+		Labels: cloneMetricLabels(labels),
+	})
+}
+
+func modelHostDiagnostics(cfg *FactoryServiceConfig, logger *zap.Logger) modelhost.Diagnostics {
+	diagnostics := modelhost.Diagnostics{}
+	if logger != nil {
+		diagnostics.Logger = newZapModelHostLogger(logger.Named("modelhost"))
+	}
+	if cfg != nil {
+		diagnostics.Metrics = newModelHostMetricsRecorder(cfg.InvocationMetricsRecorder)
+	}
+	return diagnostics
+}
